# Desarrollo/docs/openapi.yaml
openapi: 3.0.0

info:
  title: MiniBlog API
  version: 1.0.0
  description: API REST para gestionar autores y publicaciones de MiniBlog.

servers:
  - url: http://localhost:3000
    description: Servidor local

security: []

paths:
  /authors:
    get:
      summary: Obtener todos los autores
      responses:
        "200":
          description: Lista de autores obtenida correctamente
        "500":
          description: Error al obtener los autores

    post:
      summary: Crear un nuevo autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                bio:
                  type: string
      responses:
        "201":
          description: Autor creado correctamente
        "400":
          description: Datos inválidos o email ya registrado
        "500":
          description: Error al crear el autor

  /authors/{id}:
    get:
      summary: Obtener un autor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Autor encontrado
        "404":
          description: Autor no encontrado
        "500":
          description: Error al obtener el autor

    put:
      summary: Actualizar un autor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                bio:
                  type: string
      responses:
        "200":
          description: Autor actualizado correctamente
        "400":
          description: Datos inválidos o email ya registrado
        "404":
          description: Autor no encontrado
        "500":
          description: Error al actualizar el autor

    delete:
      summary: Eliminar un autor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Autor eliminado correctamente
        "404":
          description: Autor no encontrado
        "500":
          description: Error al eliminar el autor

  /posts:
    get:
      summary: Obtener todos los posts
      responses:
        "200":
          description: Lista de posts obtenida correctamente
        "500":
          description: Error al obtener los posts

    post:
      summary: Crear un nuevo post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - author_id
                - title
                - content
              properties:
                author_id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                published:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Post creado correctamente
        "400":
          description: Datos inválidos o autor inexistente
        "500":
          description: Error al crear el post

  /posts/{id}:
    get:
      summary: Obtener un post por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post encontrado
        "404":
          description: Post no encontrado
        "500":
          description: Error al obtener el post

    put:
      summary: Actualizar un post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - author_id
                - title
                - content
              properties:
                author_id:
                  type: integer
                title:
                  type: string
                content:
                  type: string
                published:
                  type: boolean
                  default: false
      responses:
        "200":
          description: Post actualizado correctamente
        "400":
          description: Datos inválidos o autor inexistente
        "404":
          description: Post no encontrado
        "500":
          description: Error al actualizar el post

    delete:
      summary: Eliminar un post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post eliminado correctamente
        "404":
          description: Post no encontrado
        "500":
          description: Error al eliminar el post

  /posts/author/{authorId}:
    get:
      summary: Obtener los posts de un autor con sus datos
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Posts del autor obtenidos correctamente
        "404":
          description: No se encontraron posts para este autor
        "500":
          description: Error al obtener los posts del autor